# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AwardCategory": {
        "enum": [
          "competition",
          "research",
          "innovation",
          "honor",
          "other"
        ],
        "title": "AwardCategory",
        "type": "string"
      },
...
      "AwardLevel": {
        "enum": [
          "national",
          "provincial",
          "municipal",
          "university",
          "other"
        ],
        "title": "AwardLevel",
        "type": "string"
      },
      "AwardPublic": {
        "properties": {
          "award_date": {
            "format": "date",
            "title": "Award Date",
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/AwardCategory"
          },
          "certificate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaPublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "certificate_media_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Certificate Media Id"
          },
          "competition_name": {
            "title": "Competition Name",
            "type": "string"
          },
          "cover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaPublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "cover_media_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Media Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_featured": {
            "title": "Is Featured",
            "type": "boolean"
          },
          "issuer": {
            "title": "Issuer",
            "type": "string"
          },
          "level": {
            "$ref": "#/components/schemas/AwardLevel"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "category",
          "level",
          "issuer",
          "competition_name",
          "description",
          "award_date",
          "year",
          "certificate_media_id",
          "cover_media_id",
          "certificate",
          "cover",
          "sort_order",
          "is_featured",
          "created_at",
          "updated_at"
        ],
        "title": "AwardPublic",
        "type": "object"
      },
      "AwardSort": {
        "enum": [
          "date_desc",
          "sort_order"
        ],
        "title": "AwardSort",
        "type": "string"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "NewsPublic": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "cover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaPublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "excerpt",
          "content",
          "cover",
          "published_at",
          "created_at",
          "updated_at"
        ],
        "title": "NewsPublic",
        "type": "object"
      },
...
      "PageResponse_AwardPublic_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AwardPublic"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "minimum": 0.0,
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "pages"
        ],
        "title": "PageResponse[AwardPublic]",
        "type": "object"
      },
...
      "PageResponse_GalleryItemPublic_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GalleryItemPublic"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "minimum": 0.0,
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "pages"
        ],
        "title": "PageResponse[GalleryItemPublic]",
        "type": "object"
      },
...
      "PageResponse_NewsPublic_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NewsPublic"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "minimum": 0.0,
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "pages"
        ],
        "title": "PageResponse[NewsPublic]",
        "type": "object"
      },
...
      "PageResponse_ProjectPublic_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectPublic"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "minimum": 0.0,
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "pages"
        ],
        "title": "PageResponse[ProjectPublic]",
        "type": "object"
      },
...
      "PageResponse_ResearchAreaPublic_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResearchAreaPublic"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "minimum": 0.0,
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "pages"
        ],
        "title": "PageResponse[ResearchAreaPublic]",
        "type": "object"
      },
...
      "ProjectPublic": {
        "properties": {
          "cover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaPublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "demo_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 500,
                "minLength": 1,
                "pattern": "^https?://",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demo Url"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "summary",
          "description",
          "demo_url",
          "cover",
          "sort_order",
          "published_at",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectPublic",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Content management and public API for the robotics laboratory website.",
    "title": "Robotics Laboratory CMS API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/awards": {
      "get": {
        "operationId": "list_awards_api_v1_awards_get",
        "parameters": [
          {
            "in": "query",
            "name": "featured",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Featured"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AwardSort",
              "default": "date_desc"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_AwardPublic_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List visible awards",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/awards/{award_id}": {
      "get": {
        "operationId": "get_award_api_v1_awards__award_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "award_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Award Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AwardPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Award",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/gallery": {
      "get": {
        "operationId": "list_gallery_api_v1_gallery_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_GalleryItemPublic_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List visible gallery records",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/news": {
      "get": {
        "operationId": "list_news_api_v1_news_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_NewsPublic_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List published news",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/news/{slug}": {
      "get": {
        "operationId": "get_news_api_v1_news__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get News",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/projects": {
      "get": {
        "operationId": "list_projects_api_v1_projects_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_ProjectPublic_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List published projects",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/projects/{slug}": {
      "get": {
        "operationId": "get_project_api_v1_projects__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Project",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/research-areas": {
      "get": {
        "operationId": "list_research_areas_api_v1_research_areas_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_ResearchAreaPublic_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List visible research areas",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/site-settings": {
      "get": {
        "operationId": "get_site_settings_api_v1_site_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteSettingsPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Site Settings",
        "tags": [
          "public"
        ]
      }
    },
...
  }
}